Update visualizer.py
The line from scipy.misc import imresize is being changed because the imresize function from scipy.misc has been deprecated and removed in recent versions of SciPy. This means that using imresize can lead to compatibility issues and warnings or errors when running the code with newer versions of SciPy.